make GameCycleLogic interface
