+++ model/src/test/kotlin/ru/hse/se/team9/model/logic/gamecycle/MovementTest.kt
@@ -2,7 +2,6 @@ package ru.hse.se.team9.model.logic.gamecycle
 
 import arrow.core.Right
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import ru.hse.se.team9.entities.MapObject
@@ -21,7 +20,7 @@ internal class MovementTest {
     private lateinit var startMobPosition: Position
     private var width: Int = 0
     private var height: Int = 0
-    private lateinit var gameCycleLogic: GameCycleLogic
+    private lateinit var gameCycleLogicImpl: GameCycleLogicImpl
 
     @BeforeEach
     fun init() {
@@ -39,7 +38,7 @@ internal class MovementTest {
         startMobPosition = SimpleTestMap.startMobPosition
         width = SimpleTestMap.width
         height = SimpleTestMap.height
-        gameCycleLogic = GameCycleLogic(SimpleTestMap.gameMap(), generator)
+        gameCycleLogicImpl = GameCycleLogicImpl(SimpleTestMap.gameMap(), generator)
     }
 
     @Test
@@ -50,21 +49,21 @@ internal class MovementTest {
          * ..@x.#
          * ##x#.#
          */
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Down))
-        assertEquals(startPosition, gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Right))
-        assertEquals(startPosition, gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Up))
-        assertEquals(Position(startPosition.x, startPosition.y - 1), gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Up))
-        assertEquals(Position(startPosition.x, startPosition.y - 2), gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Left))
-        assertEquals(Position(startPosition.x, startPosition.y - 2), gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Up))
-        assertEquals(Position(startPosition.x, startPosition.y - 2), gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Right))
-        assertEquals(Position(startPosition.x + 1, startPosition.y - 2), gameCycleLogic.map.heroOnMap.position)
-        assertEquals(Right(InProgress), gameCycleLogic.movePlayer(Up))
-        assertEquals(Position(startPosition.x + 1, startPosition.y - 2), gameCycleLogic.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Down))
+        assertEquals(startPosition, gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Right))
+        assertEquals(startPosition, gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Up))
+        assertEquals(Position(startPosition.x, startPosition.y - 1), gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Up))
+        assertEquals(Position(startPosition.x, startPosition.y - 2), gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Left))
+        assertEquals(Position(startPosition.x, startPosition.y - 2), gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Up))
+        assertEquals(Position(startPosition.x, startPosition.y - 2), gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Right))
+        assertEquals(Position(startPosition.x + 1, startPosition.y - 2), gameCycleLogicImpl.map.heroOnMap.position)
+        assertEquals(Right(InProgress), gameCycleLogicImpl.movePlayer(Up))
+        assertEquals(Position(startPosition.x + 1, startPosition.y - 2), gameCycleLogicImpl.map.heroOnMap.position)
     }
 }
